+++ packages/cli/src/api/collectUserEditDiffs.ts
@@ -47,11 +47,13 @@ export async function collectAndSendUserEditDiffs(
 ): Promise<boolean> {
   if (!settings.files) return false;
 
-  const { resolvedPaths, placeholderPaths, transformPaths } = settings.files;
+  const { resolvedPaths, placeholderPaths, transformPaths, transformFormats } =
+    settings.files;
   const fileMapping = createFileMapping(
     resolvedPaths,
     placeholderPaths,
     transformPaths,
+    transformFormats,
     settings.locales,
     settings.defaultLocale
   );