+++ packages/cli/src/utils/__tests__/localizeStaticImports.test.ts
@@ -82,6 +82,7 @@ describe('localizeStaticImports', () => {
           placeholderPaths: { docs: '/docs' },
           resolvedPaths: ['file1', 'file2'],
           transformPaths: {},
+          transformFormats: {},
         },
         defaultLocale: 'en',
         locales: ['en', 'ja'],
@@ -96,6 +97,7 @@ describe('localizeStaticImports', () => {
       expect(createFileMapping).toHaveBeenCalledWith(
         ['file1', 'file2'],
         { docs: '/docs' },
+        {},
         {},
         ['en', 'ja'],
         'en'