+++ packages/next/CHANGELOG.md
@@ -1,5 +1,34 @@
 # gt-next
 
+## 6.16.18
+
+### Patch Changes
+
+- Updated dependencies [[`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8)]:
+  - generaltranslation@8.2.9
+  - gt-i18n@0.8.8
+  - @generaltranslation/compiler@1.3.15
+  - gt-react@10.19.7
+  - @generaltranslation/supported-locales@2.0.67
+
+## 6.16.17
+
+### Patch Changes
+
+- Updated dependencies [[`b12d57d`](https://github.com/generaltranslation/gt/commit/b12d57dab1d5cb1f602c5ac24a702b48cda7f11e)]:
+  - generaltranslation@8.2.8
+  - @generaltranslation/compiler@1.3.14
+  - gt-i18n@0.8.7
+  - gt-react@10.19.6
+  - @generaltranslation/supported-locales@2.0.66
+
+## 6.16.16
+
+### Patch Changes
+
+- Updated dependencies [[`28d0c06`](https://github.com/generaltranslation/gt/commit/28d0c06f3e8366fc2c119b7792620c4764eda2de)]:
+  - @generaltranslation/compiler@1.3.13
+
 ## 6.16.15
 
 ### Patch Changes